AE-453: Fixing code to use traceRpcUrl and process by transaction not by block.
