+++ agent-config.json
@@ -1,6 +1,3 @@
 {
-  "monitorMintCalls": {
-    "rpcTimeoutMilliseconds": 15000
-  },
   "umaEverestId": "0x9ed51155fa709f1bc3b26b8fec03df7010177362"
 }
