+++ src/agent.js
@@ -1,4 +1,5 @@
 const deployerWatchAgent = require('./deployer-watch');
+const monitorMintCallsAgent = require('./monitor-mint-calls');
 
 const handleTransaction = async (txEvent) => {
   const findings = [];
@@ -7,7 +8,13 @@ const handleTransaction = async (txEvent) => {
     deployerWatchAgent.handleTransaction(txEvent),
   ]);
 
+  const [monitorMintCallsFindings] = await Promise.all([
+    monitorMintCallsAgent.handleTransaction(txEvent),
+  ]);
+
   findings.push(...deployerWatchFindings);
+  findings.push(...monitorMintCallsFindings);
+
   return findings;
 };
